HDDS-10046. Replace PrecomputedVolumeSpace with SpaceUsageSource.Fixed
